refactored screens
