+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:starthub_mobile_pjt/providers/user.dart';
 import 'package:starthub_mobile_pjt/screen/project-detail.dart';
-// import 'package:starthub_mobile_pjt/authentication/sign_up.dart';
+// import 'package:starthub_mobile_pjt/authentication/signup_screen.dart';
 // import 'package:starthub_mobile_pjt/screen/signup_screen.dart';
 import 'package:starthub_mobile_pjt/screen/project_screen.dart';
