+++ lib/screen/login_screen.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 import '../constants.dart';
-import '../presets.dart';
-import 'sign_up.dart';
+import '../widget/presets.dart';
+import 'signup_screen.dart';
 
 class LogIn extends StatefulWidget {
   @override
